+++ src/obr/init.ts
@@ -1,14 +1,16 @@
-import OBR, { isShape, isText, Item, Metadata } from "@owlbear-rodeo/sdk";
+import OBR, { Item, Text, Metadata } from "@owlbear-rodeo/sdk";
 import { ID, characterMetadata, sceneMetadata } from "../helper/variables.ts";
-import { prepareDisplayChanges } from "../helper/infoHelpers.ts";
 import { migrate102To103 } from "../migrations/v103.ts";
 import { migrate105To106 } from "../migrations/v106.ts";
 import { compare } from "compare-versions";
-import { HpTrackerMetadata, SceneMetadata } from "../helper/types.ts";
+import { HpTrackerMetadata, SceneMetadata, TextItemChanges } from "../helper/types.ts";
 import { migrate111To112 } from "../migrations/v112.ts";
 import { migrate112To113 } from "../migrations/v113.ts";
+import { updateHpBar } from "../helper/shapeHelpers.ts";
+import { handleTextVisibility, updateText, updateTextChanges } from "../helper/textHelpers.ts";
+import { getAttachedItems } from "../helper/helpers.ts";
 
-const version = "1.1.3";
+const version = "1.2.0";
 
 /**
  * All character items get the default values for the HpTrackeMetadata.
@@ -20,7 +22,7 @@ const initItems = async () => {
             if (!(characterMetadata in item.metadata)) {
                 // initializing a variable gives us type safety
                 const initialMetadata: HpTrackerMetadata = {
-                    name: "",
+                    name: item.name,
                     hp: 0,
                     maxHp: 0,
                     armorClass: 0,
@@ -37,83 +39,7 @@ const initItems = async () => {
         });
     };
 
-    await OBR.scene.items.updateItems((item) => item.layer === "CHARACTER", addMetaData);
-};
-
-/**
- * The Texts that display the current HP of a Character Item must be updated anytime the metadata of the Character Items
- * is changed.
- *
- */
-const initLocalItems = async () => {
-    const role = await OBR.player.getRole();
-
-    const updateScene = async (items: Item[]) => {
-        const characters = items.filter((item) => item.layer === "CHARACTER");
-        const changes = await prepareDisplayChanges(characters, role);
-        if (changes.textItems.size > 0) {
-            await OBR.scene.local.updateItems(isText, (texts) => {
-                texts.forEach((text) => {
-                    if (changes.textItems.has(text.id)) {
-                        const change = changes.textItems.get(text.id);
-                        if (change) {
-                            if (change.text) {
-                                text.text.plainText = change.text;
-                            }
-                            if (change.visible !== undefined) {
-                                text.visible = change.visible;
-                            }
-                            if (change.position) {
-                                text.position = change.position;
-                            }
-                        }
-                    }
-                });
-            });
-        }
-        if (changes.shapeItems.size > 0) {
-            await OBR.scene.local.updateItems(isShape, (shapes) => {
-                shapes.forEach((shape) => {
-                    if (changes.shapeItems.has(shape.id)) {
-                        const change = changes.shapeItems.get(shape.id);
-                        if (change) {
-                            if (change.width) {
-                                shape.width = change.width;
-                            }
-                            if (change.visible !== undefined) {
-                                shape.visible = change.visible;
-                            }
-                            if (change.position) {
-                                shape.position = change.position;
-                            }
-                            if (change.color) {
-                                shape.style.fillColor = change.color;
-                            }
-                        }
-                    }
-                });
-            });
-        }
-    };
-
-    const sceneItems = await OBR.scene.items.getItems(
-        (item) => item.layer === "CHARACTER" && characterMetadata in item.metadata
-    );
-    await updateScene(sceneItems);
-
-    // Triggers everytime any item is changed
-    OBR.scene.items.onChange(async (items) => {
-        // But we only care about Character Items
-        await updateScene(items);
-    });
-
-    // Triggers when the scene metadata is changed
-    OBR.scene.onMetadataChange(async () => {
-        const items = await OBR.scene.items.getItems(
-            (item) => item.layer === "CHARACTER" && characterMetadata in item.metadata
-        );
-        await updateScene(items);
-    });
+    await OBR.scene.items.updateItems((item) => item.layer === "CHARACTER" || item.layer === "MOUNT", addMetaData);
 };
 
 const initScene = async () => {
@@ -148,6 +74,7 @@ const setupContextMenu = async () => {
                 filter: {
                     every: [
                         { key: "layer", value: "CHARACTER" },
+                        { key: "layer", value: "MOUNT", coordinator: "||" },
                         {
                             key: ["metadata", `${characterMetadata}`, "hpTrackerActive"],
                             value: true,
@@ -164,6 +91,10 @@ const setupContextMenu = async () => {
                         const metadata = item.metadata[characterMetadata] as HpTrackerMetadata;
                         metadata.hp = Math.min(metadata.hp + 1, metadata.maxHp);
                         item.metadata[characterMetadata] = { ...metadata };
+                        updateHpBar(metadata.hpBar, item.id, { ...metadata });
+                        updateText(metadata.hpOnMap || metadata.acOnMap, metadata.canPlayersSee, item.id, {
+                            ...metadata,
+                        });
                     }
                 });
             });
@@ -178,6 +109,7 @@ const setupContextMenu = async () => {
                 filter: {
                     every: [
                         { key: "layer", value: "CHARACTER" },
+                        { key: "layer", value: "MOUNT", coordinator: "||" },
                         {
                             key: ["metadata", `${characterMetadata}`, "hpTrackerActive"],
                             value: true,
@@ -200,6 +132,10 @@ const setupContextMenu = async () => {
                         const metadata = item.metadata[characterMetadata] as HpTrackerMetadata;
                         metadata.hp = allowNegativeNumbers ? metadata.hp - 1 : Math.max(metadata.hp - 1, 0);
                         item.metadata[characterMetadata] = { ...metadata };
+                        updateHpBar(metadata.hpBar, item.id, { ...metadata });
+                        updateText(metadata.hpOnMap || metadata.acOnMap, metadata.canPlayersSee, item.id, {
+                            ...metadata,
+                        });
                     }
                 });
             });
@@ -214,6 +150,7 @@ const setupContextMenu = async () => {
                 filter: {
                     every: [
                         { key: "layer", value: "CHARACTER" },
+                        { key: "layer", value: "MOUNT", coordinator: "||" },
                         { key: ["metadata", `${characterMetadata}`], value: undefined, coordinator: "||" },
                         {
                             key: ["metadata", `${characterMetadata}`, "hpTrackerActive"],
@@ -228,7 +165,7 @@ const setupContextMenu = async () => {
                 icon: "/iconOff.svg",
                 label: "HP Tracker",
                 filter: {
-                    every: [{ key: "layer", value: "CHARACTER" }],
+                    every: [{ key: ["metadata", `${characterMetadata}`, "hpTrackerActive"], value: true }],
                     roles: ["GM"],
                 },
             },
@@ -285,20 +222,23 @@ const migrations = async () => {
     }
 };
 
-const delay = (ms: number) => {
-    return new Promise((resolve) => setTimeout(resolve, ms));
+const sceneReady = async () => {
+    await migrations();
+    await initItems();
+    await initScene();
 };
 
-const initLocalLoop = async () => {
-    let initialized = false;
-    while (!initialized) {
-        try {
-            await initLocalItems();
-            initialized = true;
-        } catch {
-            await delay(1000);
+const registerEvents = () => {
+    // Triggers everytime any item is changed
+    OBR.scene.items.onChange(async (items) => {
+        const changeMap = new Map<string, TextItemChanges>();
+        const tokens = items.filter((item) => characterMetadata in item.metadata);
+        for (const token of tokens) {
+            const texts = await getAttachedItems(token.id, "TEXT");
+            await handleTextVisibility(token, texts as Array<Text>, changeMap);
         }
-    }
+        await updateTextChanges(changeMap);
+    });
 };
 
 OBR.onReady(async () => {
@@ -306,12 +246,14 @@ OBR.onReady(async () => {
     await setupContextMenu();
     OBR.scene.onReadyChange(async (isReady) => {
         if (isReady) {
-            await migrations();
-            await initItems();
-            await initScene();
+            await sceneReady();
         }
     });
-    try {
-        await initLocalLoop();
-    } catch {}
+
+    const isReady = await OBR.scene.isReady();
+    if (isReady) {
+        await sceneReady();
+    }
+
+    registerEvents();
 });